backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
sqlalchemy
pymysql
python-dotenv
python-multipart
python-dateutil
python-json-logger
bcrypt
PyJWT
passlib[bcrypt]
cryptography
